pyproject.toml
[project]
name = "python-package-newsletter"
version = "0.1.0"
description = "Automated Python package update newsletter"
authors = [{name = "Automation"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "requests>=2.31.0",
    "PyYAML>=6.0",
    "packaging>=23.2"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
